refactor: Move files to introduce feature to the presentation layer
