+++ lib/presentation/common/components/error_dialog.dart
