+++ lib/presentation/common/mixins/on_layout_built_mixin.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_training/presentation/screens/initial_screen.dart';
+import 'package:flutter_training/presentation/common/screens/initial_screen.dart';
 
 mixin OnLayoutBuiltMixin on State<InitialScreen> {
   @override
