+++ lib/presentation/common/screens/initial_screen.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_training/application/usecases/reload_weather_usecase.dart';
-import 'package:flutter_training/presentation/mixins/on_layout_built_mixin.dart';
-import 'package:flutter_training/presentation/screens/weather_screen.dart';
+import 'package:flutter_training/application/weather/usecases/reload_weather_usecase.dart';
+import 'package:flutter_training/presentation/common/mixins/on_layout_built_mixin.dart';
+import 'package:flutter_training/presentation/weather/screens/weather_screen.dart';
 
 class InitialScreen extends StatefulWidget {
   const InitialScreen({
